requirements.txt
pandas
PyYAML
SQLAlchemy
psycopg2-binary
fastparquet
requests
pandas 
tabulate
